+++ lua/github-theme/palette/github_light_default.lua
@@ -13,10 +13,7 @@ local scale = primitives.scale
 C.WHITE = C(scale.white)
 C.BLACK = C(scale.black)
 C.BG = C(assert(primitives.canvas.default))
-
-local function alpha(color, a)
-  return color:alpha_blend(a):to_css()
-end
+local BG = C(assert(primitives.canvas.default))
 
 -- Temp override until Primitives are updated
 local palette = {
@@ -47,76 +44,76 @@ local palette = {
   border = {
     default = scale.gray[3],
     muted = C(scale.gray[3]):lighten(0.03), -- TODO: lighten method not working
-    subtle = alpha(C(scale.black), 0.15),
+    subtle = BG:blend(C(scale.black), 0.15):to_css(),
   },
 
   neutral = {
     emphasis_plus = scale.gray[10],
     emphasis = scale.gray[6],
-    muted = alpha(C(scale.gray[4]), 0.2),
-    subtle = alpha(C(scale.gray[2]), 0.5),
+    muted = BG:blend(C(scale.gray[4]), 0.2):to_css(),
+    subtle = BG:blend(C(scale.gray[2]), 0.5):to_css(),
   },
 
   accent = {
     fg = scale.blue[6],
     emphasis = scale.blue[6],
-    muted = alpha(C(scale.blue[4]), 0.4),
+    muted = BG:blend(C(scale.blue[4]), 0.4):to_css(),
     subtle = scale.blue[1],
   },
 
   success = {
     fg = scale.green[6],
     emphasis = '#1f883d',
-    muted = alpha(C(scale.green[4]), 0.4),
+    muted = BG:blend(C(scale.green[4]), 0.4):to_css(),
     subtle = scale.green[1],
   },
 
   attention = {
     fg = scale.yellow[6],
     emphasis = scale.yellow[6],
-    muted = alpha(C(scale.yellow[4]), 0.4),
+    muted = BG:blend(C(scale.yellow[4]), 0.4):to_css(),
     subtle = scale.yellow[1],
   },
 
   severe = {
     fg = scale.orange[6],
     emphasis = scale.orange[6],
-    muted = alpha(C(scale.orange[4]), 0.4),
+    muted = BG:blend(C(scale.orange[4]), 0.4):to_css(),
     subtle = scale.orange[1],
   },
 
   danger = {
     fg = '#d1242f',
     emphasis = scale.red[6],
-    muted = alpha(C(scale.red[4]), 0.4),
+    muted = BG:blend(C(scale.red[4]), 0.4):to_css(),
     subtle = scale.red[1],
   },
 
   open = {
     fg = scale.green[6],
     emphasis = '#1f883d',
-    muted = alpha(C(scale.green[4]), 0.4),
+    muted = BG:blend(C(scale.green[4]), 0.4):to_css(),
     subtle = scale.green[1],
   },
 
   closed = {
     fg = '#d1242f',
     emphasis = scale.red[6],
-    muted = alpha(C(scale.red[4]), 0.4),
+    muted = BG:blend(C(scale.red[4]), 0.4):to_css(),
     subtle = scale.red[1],
   },
 
   done = {
     fg = scale.purple[6],
     emphasis = scale.purple[6],
-    muted = alpha(C(scale.purple[4]), 0.4),
+    muted = BG:blend(C(scale.purple[4]), 0.4):to_css(),
     subtle = scale.purple[1],
   },
 
   sponsors = {
     fg = scale.pink[6],
     emphasis = scale.pink[6],
-    muted = alpha(C(scale.pink[4]), 0.4),
+    muted = BG:blend(C(scale.pink[4]), 0.4):to_css(),
     subtle = scale.pink[1],
   },
 }
@@ -126,8 +123,8 @@ local function generate_spec(pal)
   local spec = {
     bg0  = pal.canvas.inset,                            -- Dark bg (popup and float)
     bg1  = pal.canvas.default,                          -- Default bg
-    bg2  = alpha(C(pal.neutral.emphasis), 0.1),         -- Lighter bg (colorcolumn Folds)
-    bg3  = alpha(C(pal.scale.blue[9]), 0.1),            -- Lighter bg (cursor line)
+    bg2  = BG:blend(C(pal.neutral.emphasis), 0.1):to_css(),      -- Lighter bg (colorcolumn Folds)
+    bg3  = BG:blend(C(pal.scale.blue[9]), 0.1):to_css(),         -- Lighter bg (cursor line)
     bg4  = pal.scale.gray[6],                           -- Conceal
 
     fg0  = pal.fg.subtle,                               -- Lighter fg
@@ -135,9 +132,9 @@ local function generate_spec(pal)
     fg2  = pal.scale.gray[9],                           -- Darker fg (status line)
     fg3  = pal.scale.gray[7],                           -- Darker fg (line numbers, fold columns)
 
-    sel0 = alpha(C(pal.accent.fg), 0.15),               -- Visual selection bg
-    sel1 = alpha(C(pal.accent.muted), 0.90),            -- Popup sel bg
-    sel2 = alpha(C(pal.attention.emphasis), 0.3)        -- Search bg
+    sel0 = BG:blend(C(pal.accent.fg), 0.15):to_css(),            -- Visual selection bg
+    sel1 = BG:blend(C(pal.accent.muted), 0.90):to_css(),         -- Popup sel bg
+    sel2 = BG:blend(C(pal.attention.emphasis), 0.3):to_css()     -- Search bg
   }
 
   spec.syntax = {
@@ -173,16 +170,16 @@ local function generate_spec(pal)
   }
 
   spec.diag_bg = {
-    error = C(spec.bg1):blend(C(spec.diag.error), 0.15):to_css(),
-    warn  = C(spec.bg1):blend(C(spec.diag.warn), 0.15):to_css(),
-    info  = C(spec.bg1):blend(C(spec.diag.info), 0.15):to_css(),
-    hint  = C(spec.bg1):blend(C(spec.diag.hint), 0.15):to_css(),
+    error = BG:blend(C(spec.diag.error), 0.15):to_css(),
+    warn  = BG:blend(C(spec.diag.warn), 0.15):to_css(),
+    info  = BG:blend(C(spec.diag.info), 0.15):to_css(),
+    hint  = BG:blend(C(spec.diag.hint), 0.15):to_css(),
   }
 
   spec.diff = {
-    add    = alpha(C(pal.scale.green[7]), 0.3),
-    delete = alpha(C(pal.scale.red[7]), 0.3),
-    change = alpha(C(pal.scale.yellow[7]), 0.3),
+    add    = BG:blend(C(pal.scale.green[7]), 0.3):to_css(),
+    delete = BG:blend(C(pal.scale.red[7]), 0.3):to_css(),
+    change = BG:blend(C(pal.scale.yellow[7]), 0.3):to_css(),
     text   = spec.fg0
   }